refactor(ksc): split static pose + loop-schedule families into ParsekKSC.Playback.cs
Relocate the static KSC pose-resolution family (TryInterpolateKscPlaybackPose
+ frame selection / track-section helpers + point/segment/relative pose
resolvers + KscAnchorFrame/KscPoseResolution structs + lookup delegates +
KscFlatPointFrameSourceKey const) and the static loop-schedule family
(TryGetLoopSchedule / TryComputeLoopUT / GetLoopIntervalSeconds overloads)
verbatim into a new partial-class file. Zero logic change: methods moved as-is,
all call sites resolve across the partial. Add the partial keyword to the class
declaration and allowlist the new file for the ERS/ELS grep-audit gate
(it carries the [ERS-exempt] CommittedRecordings read inside TryGetLoopSchedule).
